requirements.txt
python-telegram-bot==11.1.0
tinydb==3.12.2
tinydb-serialization==1.0.4